+++ NEWS.md
@@ -1,7 +1,9 @@
 # docorator development version
-
+Improvements
 * Add `qmd` PDF engine
 
+Bug fixes
+* Fix issue where `fancyrow()` did not reject string vectors of length > 1 as arguments
 # docorator 0.4.0
 
 First open source release